Revert "ChainSpec: all and everything at once" (#16502)
Reverts erigontech/erigon#15974
due to testing calls to unregistered chains

+++ cmd/observer/reports/clients_estimate_report.go
@@ -35,7 +35,13 @@ type ClientsEstimateReport struct {
 	Clients []ClientsEstimateReportEntry
 }
 
-func CreateClientsEstimateReport(ctx context.Context, db database.DB, limit uint, maxPingTries uint, networkID uint64) (*ClientsEstimateReport, error) {
+func CreateClientsEstimateReport(
+	ctx context.Context,
+	db database.DB,
+	limit uint,
+	maxPingTries uint,
+	networkID uint,
+) (*ClientsEstimateReport, error) {
 	clientsReport, err := CreateClientsReport(ctx, db, limit, maxPingTries, networkID)
 	if err != nil {
 		return nil, err
@@ -49,7 +55,7 @@ func CreateClientsEstimateReport(ctx context.Context, db database.DB, limit uint
 		}
 		clientName := topClient.Name
 
-		sameNetworkCount, err := db.CountClients(ctx, clientName+"/", maxPingTries, uint(networkID))
+		sameNetworkCount, err := db.CountClients(ctx, clientName+"/", maxPingTries, networkID)
 		if err != nil {
 			return nil, err
 		}
